minor layout changes
