+++ potato/src/candidatesApp/App.js
@@ -8,28 +8,24 @@ import {createStore, applyMiddleware} from 'redux';
 import reducers from './reducers'
 import ReduxThunk from 'redux-thunk'
 import RegisterForm from './components/RegisterForm';
-import CandidatesForm from './components/CandidatesForm'
 import {StackNavigator} from 'react-navigation'
 import CandidatesList from './components/CandidatesList'
-import CandidatesListTrial from './components/CandidatesListTrial'
-import CandidatesFormREDUX from './components/CandidatesFormREDUX'
+import CandidatesForm from './components/CandidatesForm'
 
 const BasicApp = StackNavigator({
-  Register: {screen: RegisterForm},
-  CandidatesListTrial: {screen: CandidatesListTrial},
-  CandidatesFormREDUX: {screen: CandidatesFormREDUX},
+  Register      : {screen: RegisterForm},
   CandidatesList: {screen: CandidatesList},
   CandidatesForm: {screen: CandidatesForm},
 });
 
 class App extends Component {
-    componentWillMount() {
+  componentWillMount() {
     const config = {
-      apiKey: "AIzaSyDZYlySNVAPsym5GCTMCgE4x_p7GsFpWKo",
-      authDomain: "potato-5211a.firebaseapp.com",
-      databaseURL: "https://potato-5211a.firebaseio.com",
-      projectId: "potato-5211a",
-      storageBucket: "potato-5211a.appspot.com",
+      apiKey           : "AIzaSyDZYlySNVAPsym5GCTMCgE4x_p7GsFpWKo",
+      authDomain       : "potato-5211a.firebaseapp.com",
+      databaseURL      : "https://potato-5211a.firebaseio.com",
+      projectId        : "potato-5211a",
+      storageBucket    : "potato-5211a.appspot.com",
       messagingSenderId: "397918687238"
     };
     firebase.initializeApp(config);
@@ -40,7 +36,7 @@ class App extends Component {
 
     return (
       <Provider store={store}>
-          <BasicApp/>
+        <BasicApp/>
       </Provider>
     )
   }
